Remove default dtype arguments, require explicit dtype everywhere
Following the principle "DEFAULT ARGUMENTS ARE EVIL":
- make_rope_periods: dtype is now required (positional)
- grid_offsets: dtype is now required
- make_grid_positions: dtype is now required
- glimpse_positions: dtype is now required
- IJEPABackbone: rope_dtype is now required

All call sites updated to pass dtype explicitly using kwargs for clarity.
This makes the data flow explicit and prevents silent type mismatches.

+++ avp_vit/backbone/ijepa/test.py
@@ -13,7 +13,7 @@ def test_forward_matches_native():
     native = vit_small(img_size=[112], patch_size=16)
     native.eval()
 
-    backbone = IJEPABackbone(native)
+    backbone = IJEPABackbone(native, rope_dtype=torch.float32)
 
     B = 2
     img = torch.randn(B, 3, 112, 112)
@@ -34,7 +34,7 @@ def test_forward_matches_native():
 def test_properties():
     """Backbone exposes correct properties."""
     native = vit_small(img_size=[112], patch_size=16)
-    backbone = IJEPABackbone(native)
+    backbone = IJEPABackbone(native, rope_dtype=torch.float32)
 
     assert backbone.embed_dim == 384
     assert backbone.num_heads == 6
